kdumpctl: improve do_estimate by considering unpacked initramfs size
A memory peak in kdump occurs when the initramfs is being decompressed. At this
moment, both the compressed and decompressed initramfs coexist in memory. After
this, the compressed initramfs is released.

This patch enhances the `do_estimate` function in `kdumpctl` by incorporating
the size of the unpacked initramfs contents into the crashkernel memory
estimation. The following changes have been made:

- Added a helper function to extract and measure the size of the unpacked
  initramfs, including any squashfs/erofs images.
- Adjusted the default runtime reservation from 64M to 32M to better reflect
  typical use cases.
- Updated output to display the unpacked initramfs size.

These improvements help ensure a more accurate memory reservation for kdump
environments.

Signed-off-by: Lichen Liu <lichliu@redhat.com>