kdumpctl: Enhance get_grub_kernel_boot_parameter for multiple grub entries for same kernel
When there are multiple grub entries for the same kernel, get_grub_kernel_boot_parameter()
uses the first matching $_para value, which may cause inconsistent behavior in the function
that calls it.

When calling this function, there are several cases:

    1. The kernel has only one grub entry. This is the scenario we assumed earlier, return
    the value of $_para or "" if it does not exist.

    2. The kernel has multiple grub entries, and

        1) some entries are missing $_para, or they contain different values. Print a
	warning message and return "inconsistent_value". All matched grub entries will be
	updated when calling `kdumpctl reset-crashkernel` in this case.

        2) All entries have the same $_para, return the unique value.

After this patch, this function will return "", "inconsistent_value" or a unique value.

Signed-off-by: Lichen Liu <lichliu@redhat.com>